Resolve different signedness warning
The size() by default returns size_t datatype. Update the data type for
the loop iterator to resolve compiler warning for signedness mismatch.

+++ src/google/protobuf/compiler/java/internal_helpers.cc
@@ -177,7 +177,7 @@ void GenerateLarge(
         }},
        {"aliases",
         [&] {
-          for (int i = 0; i < (int)aliases.size(); i++) {
+          for (size_t i = 0; i < aliases.size(); i++) {
             WriteEnumValueDocComment(printer, aliases[i].first,
                                      context->options());
             printer->Emit({{"name", aliases[i].first->name()},
